requirements.txt
cachetools==4.2.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
et-xmlfile==1.1.0
Flask==2.0.2
google==3.0.0
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.28.0
google-cloud-bigquery-storage==2.9.0
google-cloud-core==2.0.0
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.21
lxml==4.6.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
pandas-gbq==0.15.0
proto-plus==1.19.2
protobuf==3.18.0
psycopg2-binary==2.9.1
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.19
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.1
